% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_disc.R
\name{contr.ltfr}
\alias{contr.ltfr}
\title{Wrapper for \code{\link[caret]{contr.ltfr}}}
\usage{
contr.ltfr(...)
}
\arguments{
\item{...}{\code{\link[caret]{contr.ltfr}} parameters.}
}
\description{
Simply redirects the call to \code{\link[caret]{contr.ltfr}}, this is done
to avoid errors when the caret namespace is not attached.
}
\examples{
contr.ltfr(2)
}
\keyword{internal}
