% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_cutpoints.R
\name{cutpoints}
\alias{cutpoints}
\alias{cutpoints,cutpoints-method}
\title{Method for obtaining the cutpoints and / or regroupments of a discretization of class \code{\link{glmdisc}}.}
\description{
This defines the method "cutpoints" which will output the cutpoints found by the S4 class \code{\link{glmdisc}}.
}
