% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.link.R
\name{discretize_link}
\alias{discretize_link}
\title{Discretize a dataset using a trained SEM discretization scheme}
\usage{
discretize_link(link, df)
}
\arguments{
\item{link}{A multinomial logit model.}

\item{df}{The dataframe, containing the same variables as the one used to train the discretization scheme, to be discretized.}
}
\description{
This function discretizes a user-specified dataset using a pre-trained SEM discretization scheme.
}
\references{
Asad Hasan, Wang Zhiyu and Alireza S. Mahani (2015). mnlogit: Multinomial Logit Model. R package version 1.2.4. \url{https://CRAN.R-project.org/package=mnlogit}
}
\keyword{discretization,}
\keyword{predict}
