% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{glmdisc-class}
\alias{glmdisc-class}
\title{Class glmdisc}
\description{
Class \code{glmdisc} represents a discretization scheme associated with its optimal logistic regression model.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{The parameters associated with the method.}

\item{\code{best.disc}}{The best discretization scheme found by the method given its parameters.}

\item{\code{performance}}{The performance obtained with the method given its parameters.}

\item{\code{disc.data}}{The discretized data: test set if test is TRUE; if test is FALSE and validation is TRUE, then it provides the discretized validation set. Otherwise, it provides the discretized training set.}

\item{\code{disc.data}}{The continuous data: test set if test is TRUE; if test is FALSE and validation is TRUE, then it provides the discretized validation set. Otherwise, it provides the discretized training set.}
}}

