% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmdisc.R
\docType{package}
\name{glmdisc-package}
\alias{glmdisc-package}
\title{glmdisc: A package for discretizing continuous features, grouping categorical features' values and optimizing it for logistic regression.}
\description{
The glmdisc package provides two important functions:
glmdisc and its associate method discretize.
}
\section{\code{\link{glmdisc}} function}{

The glmdisc function ...
}

\section{\code{\link{discretize}} function}{

The discretize function ...
}

\author{
Adrien Ehrhardt, Vincent Vandewalle, Christophe Biernacki, Philippe Heinrich.
}
