% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_plot.R
\name{plot}
\alias{plot}
\alias{plot,glmdisc-method}
\title{Method for predicting on a new input dataset given a discretization scheme and its associated model of class \code{\link{glmdisc}}.}
\description{
This defines the method "plot" which will plot some useful graphs for the discretization scheme of S4 class \code{\link{glmdisc}}
}
