% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot.glmdisc}
\alias{plot,glmdisc,ANY-method}
\alias{plot,glmdisc,missing-method}
\title{Plots for the discretized / grouped data.}
\usage{
plot.glmdisc(x)
}
\arguments{
\item{x}{The S4 \code{\link{glmdisc}} object to plot.}
}
\description{
This defines the \code{\link{plot}} method which will plot some useful graphs for the discretization scheme of S4 class \code{\link{glmdisc}}
}
\keyword{discretization,}
\keyword{plot}
\keyword{test,}
