% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_disc.R
\name{predict}
\alias{predict}
\alias{predict,glmdisc-method}
\title{Method for predicting on a new input dataset given a discretization scheme and its associated model of class \code{\link{glmdisc}}.}
\description{
This defines the method "discretize" which will discretize a new input dataset given a discretization scheme of S4 class \code{\link{glmdisc}}
}
