% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_predict.R
\name{predict}
\alias{predict}
\alias{predict,glmdisc-method}
\title{Prediction on a raw test set of the best logistic regression model on discretized data.}
\arguments{
\item{object}{The S4 discretization object.}

\item{predictors}{The test dataframe to discretize and for which we wish to have predictions.}
}
\description{
This defines the method "discretize" which will discretize a new input dataset given a discretization scheme of S4 class \code{\link{glmdisc}}
}
\details{
This function discretizes a user-provided test dataset given a discretization scheme provided by an S4 "glmdisc" object.
It then applies the learnt logistic regression model and outputs its prediction (see \code{\link{predict.glm}}).
}
\keyword{discretization,}
\keyword{predict,}
\keyword{prediction}
\keyword{test,}
