% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictlogisticRegression}
\alias{predictlogisticRegression}
\title{Leading NA}
\usage{
predictlogisticRegression(test, parameters)
}
\arguments{
\item{test}{A matrix containing test data}

\item{parameters}{A vector containing the logistic regression parameters}
}
\description{
This function returns a logical vector identifying if 
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}
