% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_discretize.R, R/method_discretize.R
\name{discretize}
\alias{discretize}
\alias{discretize,glmdisc-method}
\title{Prediction on a raw test set of the best logistic regression model on discretized / grouped data.}
\usage{
discretize(object, data)

\S4method{discretize}{glmdisc}(object, data)
}
\arguments{
\item{object}{glmdisc object}

\item{data}{the data to discretize according to the provided discretization scheme}
}
\description{
This function discretizes a user-provided test dataset given a discretization scheme provided by an S4 \code{\link{glmdisc}} object.
It then applies the learnt logistic regression model and outputs its prediction (see \code{\link{predict.glm}}).

This defines the method "discretize" which will discretize a new input dataset given a discretization scheme of S4 class \code{\link{glmdisc}}
}
\examples{
# Simulation of a discretized logit model
set.seed(1)
x <- matrix(runif(300), nrow = 100, ncol = 3)
cuts <- seq(0, 1, length.out = 4)
xd <- apply(x, 2, function(col) as.numeric(cut(col, cuts)))
theta <- t(matrix(c(0, 0, 0, 2, 2, 2, -2, -2, -2), ncol = 3, nrow = 3))
log_odd <- rowSums(t(sapply(seq_along(xd[, 1]), function(row_id) {
  sapply(
    seq_along(xd[row_id, ]),
    function(element) theta[xd[row_id, element], element]
  )
})))
y <- rbinom(100, 1, 1 / (1 + exp(-log_odd)))

sem_disc <- glmdisc(x, y,
  iter = 50, m_start = 4, test = FALSE,
  validation = FALSE, criterion = "aic"
)
discretize(sem_disc, data.frame(x))
}
\author{
Adrien Ehrhardt.
}
