% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictlogisticRegression}
\alias{predictlogisticRegression}
\title{Predicting using a logistic regression fitted with RCpp::fast_LR.}
\usage{
predictlogisticRegression(test, parameters)
}
\arguments{
\item{test}{A matrix containing test data}

\item{parameters}{A vector containing the logistic regression parameters}
}
\description{
This function returns a numeric vector containing the probability of each observation of being of class 1 given a vector of logistic
regression parameters (usually estimated through RCpp::fast_LR).
}
